requirements.txt
absl-py==2.0.0
addict==2.4.0
cachetools==4.2.4
captcha==0.4
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
humanfriendly==10.0
idna==3.4
IMcore==0.2.7
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.15.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pyparsing==3.0.9
pyreadline3==3.4.1
PySide6==6.5.1.1
PySide6-Addons==6.5.1.1
PySide6-Essentials==6.5.1.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
QtFusion==0.5.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
shiboken6==6.5.1.1
six==1.16.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
Werkzeug==3.0.1
